refactor(vuls2): move warning helper tests to the external test package
Replace the internal test file with an exported test-side WarningEntry
(two exported fields) plus mapping wrappers in export_test.go, and fold
the tests into vuls2_test.go alongside the rest.
