+++ detector/vuls2/vuls2_test.go
@@ -31,6 +31,7 @@ import (
 	vcPackageTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection/condition/criteria/criterion/versioncriterion/package"
 	vcBinaryPackageTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection/condition/criteria/criterion/versioncriterion/package/binary"
 	vcSourcePackageTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection/condition/criteria/criterion/versioncriterion/package/source"
+	warningTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection/condition/criteria/warning"
 	segmentTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection/segment"
 	ecosystemTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection/segment/ecosystem"
 	exploitTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/exploit"
@@ -15477,6 +15478,178 @@ func TestDetectCPEs(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			if err := vuls2.DetectCPEs(tt.r, tt.cpes, tt.sesh); (err != nil) != tt.wantErr {
 				t.Errorf("DetectCPEs() error = %v, wantErr %v", err, tt.wantErr)
+			}
+		})
+	}
+}
+
+func Test_collectCriteriaWarnings(t *testing.T) {
+	type call struct {
+		conds  []conditionTypes.FilteredCondition
+		source sourceTypes.SourceID
+	}
+	tests := []struct {
+		name  string
+		calls []call
+		want  []vuls2.WarningEntry
+	}{
+		{
+			name: "no warnings yields none",
+			calls: []call{{
+				conds:  []conditionTypes.FilteredCondition{{Criteria: criteriaTypes.FilteredCriteria{Operator: criteriaTypes.CriteriaOperatorTypeOR}}},
+				source: sourceTypes.RedHatOVALv2,
+			}},
+			want: nil,
+		},
+		{
+			// One entry per distinct (source, warning): duplicates collapse
+			// within a tree, across trees of the same source, and from a
+			// criterion nested a level down, while the same warning under
+			// another source stays a separate entry.
+			name: "one entry per source and warning",
+			calls: []call{
+				{
+					source: sourceTypes.RedHatOVALv2,
+					conds: []conditionTypes.FilteredCondition{
+						{
+							Criteria: criteriaTypes.FilteredCriteria{
+								Operator: criteriaTypes.CriteriaOperatorTypeOR,
+								Criterions: []criterionTypes.FilteredCriterion{{
+									Criterion: criterionTypes.Criterion{Type: criterionTypes.CriterionType("future-criterion")},
+									Warnings: []warningTypes.Warning{
+										{Kind: warningTypes.KindEmptyRange},
+										{Kind: warningTypes.KindUnevaluableRangeType, Cause: "future-range-b"},
+									},
+								}},
+							},
+						},
+						{
+							Criteria: criteriaTypes.FilteredCriteria{
+								Operator: criteriaTypes.CriteriaOperatorTypeOR,
+								Criterias: []criteriaTypes.FilteredCriteria{{
+									Operator: criteriaTypes.CriteriaOperatorTypeAND,
+									Criterions: []criterionTypes.FilteredCriterion{{
+										Criterion: criterionTypes.Criterion{Type: criterionTypes.CriterionType("future-criterion")},
+										Warnings: []warningTypes.Warning{
+											{Kind: warningTypes.KindUnevaluableRangeType, Cause: "future-range-b"},
+											{Kind: warningTypes.KindUnevaluableRangeType, Cause: "future-range-a"},
+										},
+									}},
+								}},
+							},
+						},
+					},
+				},
+				{
+					source: sourceTypes.RedHatCSAF,
+					conds: []conditionTypes.FilteredCondition{{
+						Criteria: criteriaTypes.FilteredCriteria{
+							Operator: criteriaTypes.CriteriaOperatorTypeOR,
+							Criterions: []criterionTypes.FilteredCriterion{{
+								Criterion: criterionTypes.Criterion{Type: criterionTypes.CriterionType("future-criterion")},
+								Warnings:  []warningTypes.Warning{{Kind: warningTypes.KindUnevaluableRangeType, Cause: "future-range-a"}},
+							}},
+						},
+					}},
+				},
+			},
+			want: []vuls2.WarningEntry{
+				{Source: sourceTypes.RedHatOVALv2, Warning: warningTypes.Warning{Kind: warningTypes.KindEmptyRange}},
+				{Source: sourceTypes.RedHatOVALv2, Warning: warningTypes.Warning{Kind: warningTypes.KindUnevaluableRangeType, Cause: "future-range-b"}},
+				{Source: sourceTypes.RedHatOVALv2, Warning: warningTypes.Warning{Kind: warningTypes.KindUnevaluableRangeType, Cause: "future-range-a"}},
+				{Source: sourceTypes.RedHatCSAF, Warning: warningTypes.Warning{Kind: warningTypes.KindUnevaluableRangeType, Cause: "future-range-a"}},
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			var entries []vuls2.WarningEntry
+			for _, c := range tt.calls {
+				for _, cond := range c.conds {
+					entries = vuls2.CollectCriteriaWarnings(cond.Criteria, c.source, entries)
+				}
+			}
+			if diff := gocmp.Diff(tt.want, entries, gocmpopts.EquateEmpty()); diff != "" {
+				t.Errorf("collectCriteriaWarnings() (-expected +got):\n%s", diff)
+			}
+		})
+	}
+}
+
+func Test_mergeWarningEntries(t *testing.T) {
+	tests := []struct {
+		name     string
+		dst, add []vuls2.WarningEntry
+		want     []vuls2.WarningEntry
+	}{
+		{
+			name: "duplicates collapse, new entries append",
+			dst: []vuls2.WarningEntry{
+				{Source: sourceTypes.RedHatOVALv2, Warning: warningTypes.Warning{Kind: warningTypes.KindEmptyRange}},
+			},
+			add: []vuls2.WarningEntry{
+				{Source: sourceTypes.RedHatOVALv2, Warning: warningTypes.Warning{Kind: warningTypes.KindEmptyRange}},
+				{Source: sourceTypes.RedHatCSAF, Warning: warningTypes.Warning{Kind: warningTypes.KindEmptyRange}},
+			},
+			want: []vuls2.WarningEntry{
+				{Source: sourceTypes.RedHatOVALv2, Warning: warningTypes.Warning{Kind: warningTypes.KindEmptyRange}},
+				{Source: sourceTypes.RedHatCSAF, Warning: warningTypes.Warning{Kind: warningTypes.KindEmptyRange}},
+			},
+		},
+		{
+			name: "empty add keeps dst",
+			dst: []vuls2.WarningEntry{
+				{Source: sourceTypes.RedHatOVALv2, Warning: warningTypes.Warning{Kind: warningTypes.KindEmptyRange}},
+			},
+			add: nil,
+			want: []vuls2.WarningEntry{
+				{Source: sourceTypes.RedHatOVALv2, Warning: warningTypes.Warning{Kind: warningTypes.KindEmptyRange}},
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if diff := gocmp.Diff(tt.want, vuls2.MergeWarningEntries(tt.dst, tt.add)); diff != "" {
+				t.Errorf("mergeWarningEntries() (-expected +got):\n%s", diff)
+			}
+		})
+	}
+}
+
+func Test_renderWarningEntries(t *testing.T) {
+	tests := []struct {
+		name    string
+		entries []vuls2.WarningEntry
+		want    []string
+	}{
+		{
+			name:    "no entries yields none",
+			entries: nil,
+			want:    []string{},
+		},
+		{
+			// An unset ("") cause and a cause-less kind both render kind-only
+			// lines; a set cause is quoted after the kind, with the source
+			// leading the parenthesized fields.
+			name: "one line per entry",
+			entries: []vuls2.WarningEntry{
+				{Source: sourceTypes.RedHatOVALv2, Warning: warningTypes.Warning{Kind: warningTypes.KindEmptyRange}},
+				{Source: sourceTypes.RedHatOVALv2, Warning: warningTypes.Warning{Kind: warningTypes.KindUnevaluablePackageType, Cause: ""}},
+				{Source: sourceTypes.RedHatOVALv2, Warning: warningTypes.Warning{Kind: warningTypes.KindUnevaluableRangeType, Cause: "future-range-a"}},
+				{Source: sourceTypes.RedHatCSAF, Warning: warningTypes.Warning{Kind: warningTypes.KindUnevaluableRangeType, Cause: "future-range-a"}},
+			},
+			want: []string{
+				`vuls2 skipped data it cannot evaluate (source: redhat-ovalv2, kind: empty-range). Detection may be incomplete; updating vuls may resolve this.`,
+				`vuls2 skipped data it cannot evaluate (source: redhat-ovalv2, kind: unevaluable-package-type). Detection may be incomplete; updating vuls may resolve this.`,
+				`vuls2 skipped data it cannot evaluate (source: redhat-ovalv2, kind: unevaluable-range-type, cause: "future-range-a"). Detection may be incomplete; updating vuls may resolve this.`,
+				`vuls2 skipped data it cannot evaluate (source: redhat-csaf, kind: unevaluable-range-type, cause: "future-range-a"). Detection may be incomplete; updating vuls may resolve this.`,
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if diff := gocmp.Diff(tt.want, vuls2.RenderWarningEntries(tt.entries)); diff != "" {
+				t.Errorf("renderWarningEntries() (-expected +got):\n%s", diff)
 			}
 		})
 	}
